Fixed connections caching Entities that are dirty and shouldn't be
Because Entitiy objects still persist after a connection clears its cache
Entities that are dirty had to ability to re-add themselves back into the cache.
When an Entity instructs a connection to cache itself the connection checks if
the cache is empty and also if another object has been created to represent the
Entity.  If either case is True then the Entity will not be allowed to cache and
SgConnection._cacheEntity() will simply return.