Added global config option DISABLE_FIELD_VALIDATE_ON_SET_VALUE
Whenever a field is set to a new value the field first makes sure that its
current value is valid.  This means when the field is set but has not previously
pulled its value from Shotgun a db query will happend so that the field can
check if the new value is the same value stored in Shotgun.  This helps in not
flagging fields as having updates when the new value is already the current
Shotgun database value.

Because some may not like this functionality it can now be disabled by setting
the config option DISABLE_FIELD_VALIDATE_ON_SET_VALUE to True and fields will
always set their value without first validating.