Fixed two bugs with Multi Entity fields value() accepting 'all' and 'default'
Multi Entity fields value() did not properly set the names of fields to sync
when 'all' was specified.  It was using the field parents fieldNames()!

When 'default' was passed the code thought it was dealing with an Entity object
however the Entity object was not yet created and the data was a dictionary.