Fixed SgFieldID and SgFieldType never having valid set to True
When the class property for valid was moved to being private the ID and Type
field classes were not updated to reflect this change so they were never
properly setting themselves as being valid.

This didn't have any impact on code except it meant that isValid() would always
return False.