SgEntity added fieldsValid() and arg "sgReturnTypes" to valuesSg()
Added new function to SgEntity fieldsValid(), returns a list of names
for fields that are valid.  Can filter return result by a list of
fields names and/or return types.

Added new arg to valuesSg() "sgReturnTypes", brings the function more
in line with others that contain a "sgFields" arg.

Also added a check in valuesSg() that returns immediately instead
of querying shotgun when the fields to return is empty.