Added support for "timecode" fields
The return type "timecode" is now supported as a field.

If the value in shotgun is greater than the number of milliseconds
in a day the field will push a warn msg to the SgField logger.

SgFieldTimeCode can not have its value set to a value greater
than the milliseconds in a day by calling SgField.setValue()