SgConnection added support for summarize SgSearchIterator added func
SgConnections now contain a function for calling the Shotgun API
summarize method.

SgSearchIterator added func for returning the current summary size
of the search filter.  It should be noted this value can change on
the fly due to database modifications so it should not be used as
an absolute value.