Added various methods for searching other Entities from the Entity class
Added SgEntity.fieldsInvalidate for invalidating fields.

Added the ability to specify which field is the name field for an Entity type.

Added SgEntity.hasNameField for querying if a name field exists.

Added SgEntity.hasMultiEntityUpdates for querying if an Entity has any multi entity field commits.

Added SgEntity.isFromSameSite for testing if two Entities belong to the same site.